Moved factory methods to helper class
