+++ TagCloud/Curves/ICurve.cs
@@ -5,13 +5,4 @@ namespace TagCloud.Curves;
 public interface ICurve
 {
     Point GetPoint(double t);
-
-    public static ICurve GetByName(string name)
-    {
-        return name.ToLower() switch
-        {
-            ArchimedeanSpiral.Name => new ArchimedeanSpiral(),
-            _ => throw new ArgumentException("There is no such algorithm")
-        };
-    }
 }
